interpolation "working"
The camera rotates all at once, the angles of rotation should also be increased by the time (angles can be represented by lines and use the funcion get point) Ex. An angle of 180 degre would be a line, in which if the percentage sended is 0.50 it will return 90, so turn 90 degrees.